Move scaling methods to DriveMessage
